Production-Editor/Source-Pull-Generator/src/requirements.txt
tk==0.1.0
selenium==4.11.2
regex==2022.10.31
docx2python==2.0.4
pandas==2.0.3
Pillow==10.0.1
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
openpyxl==3.0.10
